Added sound streaming.
It works by creating a virtual sink in pulseaudio and then routing the selected windows and the selected microphone (source) into it. Then you just have to tell discord to use the monitor of this sink as the input.

Known bugs:
If there are loopback modules loaded into pulseaudio, then the output will look buggy, when asking which windows to route, but it works.
Testing is needed.